+++ CHANGELOG.md
@@ -26,7 +26,8 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- [#68](https://github.com/zendframework/zend-feed/pull/68) no longer throw an exception for entry titles which have a string value of `0`.
+- [#68](https://github.com/zendframework/zend-feed/pull/68) updates both `Zend\Feed\Writer\AbstractFeed` and `Zend\Feed\Writer\Entry`
+  to no longer throw an exception for entry titles which have a string value of `0`.
 
 ## 2.9.0 - 2017-12-04